requirements_pip.txt
--extra-index-url https://download.pytorch.org/whl/cu117
actionlib==1.14.0
bondpy==1.8.6
catkin==0.8.10
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
dominate==2.8.0
dynamic-reconfigure==1.7.3
fonttools==4.40.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.4
importlib-resources==5.12.0
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.4
matplotlib==3.7.1
message-filters==1.16.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
requests==2.31.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
scipy==1.10.1
sensor-msgs==1.13.1
six==1.16.0
smclib==1.8.6
topic-tools==1.16.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.3.2
typing_extensions==4.6.3
urllib3==2.0.3
visdom==0.2.4
websocket-client==1.6.1
zipp==3.15.0
